Fix warning: conversion from 'unsigned int' to 'unsigned char' may change value [-Warith-conversion]

+++ link/.local/include/bitmask_array.hpp
@@ -66,14 +66,14 @@ constexpr std::array<uint8_t, 8> bitmask8_array_lowpass{
 };
 
 constexpr std::array<uint8_t, 8> bitmask8_array_highpass{
-    bitmask8_array_zero[ 0] + 1U,
-    bitmask8_array_zero[ 1] + 1U,
-    bitmask8_array_zero[ 2] + 1U,
-    bitmask8_array_zero[ 3] + 1U,
-    bitmask8_array_zero[ 4] + 1U,
-    bitmask8_array_zero[ 5] + 1U,
-    bitmask8_array_zero[ 6] + 1U,
-    bitmask8_array_zero[ 7] + 1U,
+    static_cast<uint8_t>(bitmask8_array_zero[ 0] + 1U),
+    static_cast<uint8_t>(bitmask8_array_zero[ 1] + 1U),
+    static_cast<uint8_t>(bitmask8_array_zero[ 2] + 1U),
+    static_cast<uint8_t>(bitmask8_array_zero[ 3] + 1U),
+    static_cast<uint8_t>(bitmask8_array_zero[ 4] + 1U),
+    static_cast<uint8_t>(bitmask8_array_zero[ 5] + 1U),
+    static_cast<uint8_t>(bitmask8_array_zero[ 6] + 1U),
+    static_cast<uint8_t>(bitmask8_array_zero[ 7] + 1U),
 };
 
 constexpr std::array<uint16_t, 16> bitmask16_array_one{
